book-service/pyproject.toml
[tool.poetry]
name = "book-service"
version = "0.1.0"
description = ""
authors = ["Misha4231 <mishagalko1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.136.1"
ruff = "^0.15.12"
sqlalchemy = "^2.0.49"
pydantic-settings = "^2.14.0"
uvicorn = "^0.46.0"
alembic = "^1.18.4"
psycopg2-binary = "^2.9.12"
asyncpg = "^0.31.0"
structlog = "^25.5.0"
pytest = "^9.0.3"
pytest-asyncio = "^1.3.0"
pytest-mock = "^3.15.1"
httpx = "^0.28.1"
psycopg = "^3.3.3"
anyio = "^4.13.0"
common = "^0.1.1"

[tool.poetry.group.local.dependencies]
common = { path = "../common", develop = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["..", "."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
